python-backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
llama-cpp-python==0.2.55
PyPDF2==3.0.1
python-multipart==0.0.6
